fix: 백엔드 수정사항에 따라 로그아웃/회원 탈퇴 로직 수정

+++ src/api/myPage/authAPI.ts
@@ -1,13 +1,16 @@
-// FIXME: 임시 (auth로 옮길 예정)
-
 import api from '../api';
 
+// 로그아웃 요청
 export const logout = async () => {
   const refreshToken = localStorage.getItem('refreshToken');
   if (!refreshToken) throw new Error('저장된 refreshToken이 없습니다.');
 
   try {
-    await api.post('/api/signup/logout', { refreshToken });
+    await api.delete('/api/signup/logout', {
+      headers: {
+        Authorization: `Bearer ${refreshToken}`,
+      },
+    });
     localStorage.clear(); // 토큰 및 기타 사용자 정보 제거
   } catch (error) {
     console.error('로그아웃 실패:', error);
@@ -15,10 +18,15 @@ export const logout = async () => {
   }
 };
 
+// 회원 탈퇴 요청
 export const withdraw = async () => {
   try {
-    await api.delete('/api/signup/withdraw');
-    localStorage.clear(); // 탈퇴 후 토큰 제거
+    const refreshToken = localStorage.getItem('refreshToken');
+    if (!refreshToken) throw new Error('저장된 refreshToken이 없습니다.');
+
+    await api.post('/api/signup/withdraw', { refreshToken });
+
+    localStorage.clear(); // 토큰 및 기타 사용자 정보 제거
   } catch (error) {
     console.error('회원 탈퇴 실패:', error);
     throw error;
